backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.7.1
faiss-cpu==1.9.0.post1
numpy==2.4.4
python-multipart==0.0.20
